Move error classes of extra plugins
