+++ zmon_worker_monitor/zmon_worker/errors.py
@@ -10,15 +10,6 @@ class SecurityError(Exception):
     pass
 
 
-class NagiosError(CheckError):
-    def __init__(self, output):
-        self.output = output
-        super(NagiosError, self).__init__()
-
-    def __str__(self):
-        return 'NagiosError. Command output: {}'.format(self.output)
-
-
 class InsufficientPermissionsError(CheckError):
     def __init__(self, user, entity):
         self.user = user
@@ -26,15 +17,6 @@ class InsufficientPermissionsError(CheckError):
 
     def __str__(self):
         return 'Insufficient permisions for user {} to access {}'.format(self.user, self.entity)
-
-
-class SnmpError(CheckError):
-    def __init__(self, message):
-        self.message = message
-        super(SnmpError, self).__init__()
-
-    def __str__(self):
-        return 'SNMP Error. Message: {}'.format(self.message)
 
 
 class JmxQueryError(CheckError):
